do not delegate the *nodata* parameter to endpoints
Since the *nodata* query parameter both is optional and does not have
any effects on the returning result data I will not forward the
parameter to endpoints. However, this does not solve the problem how we
should proceed with other optional FDSNWS query parameters (i.e.
especially those which are relevant for results).