[BUG][FED]: Correct eidaws-wfcatalog *comma-issue*
Appending a comma to the eidaws-wfcatalog output file stream was
evaluated incorrectly. Instead of appending a comma this fix prepends a
comma in case valid data was already streamed to the client. This
approach should be much safer i.e. even when endpoints return errors
instead of data.

Thanks to @Jollyfant for this catch.

References: #56
Closes: #56