[ENH][FED]: Remove tempfiles (dataselect)
Garbage collect temporary files created by eida-federator depending on
KeepTemfiles granularity. Implement the feature both for RawDownloadTask
and RawSplitAndAlignTask. Besides, RawSplitAndAlignTask is able to
*kill* pending requests.

The approach chosen is based on a lockable request related context. As
soon as an IOError while writing to the socket is detected the request
related context lock is released. Besides, calling
processor._terminate() performes garbage collection for ready but
pending (ready to be streamed) tasks. Tasks not ready yet consume the
lock release and perform garbage collection by themselves since such
tasks are not accessible anymore.

Instead of a simple lock file, in future the context might be
encapsulated by means of a DB, memcached etc.

The previous approach implemented event propagation by means of setting
an mp.Event(). When initializing worker pools the event was passed using
a global variable. However, globals are not threadsafe such that
concurrent requests could not be handled.
Implementing a solution based on the flask.g object wasn't possible,
neither. Even though the application context could be passed to tasks
the request context was lost.

References: #31, #67